backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gspread==5.12.0
google-auth==2.29.0
pandas==2.0.3